Minor restructure
- Move internal packages to top-level package
- Fix Makefile so it can be run from arbitrary working directory
