+++ internal/common.go
